requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.13.0
python-decouple==3.8
python-dotenv==1.1.0
Werkzeug==3.1.3
